pyproject.toml
[project]
name = "language-learning-tutor"
version = "0.1.0"
description = "AI-Powered English Learning Assistant with Modern Web UI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "functions-framework>=3.8.2",
    "google-cloud-firestore>=2.20.2",
    "google-cloud-secret-manager>=2.23.3",
    "google-cloud-speech>=2.32.0",
    "google-generativeai>=0.8.5",
    "requests>=2.32.3",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "firebase-admin>=7.4.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
    "rust-just>=1.41.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app_core",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
